修改缓存设计

+++ common/src/main/java/com/gproject/common/utils/date/DateUtils.java
@@ -37,27 +37,27 @@ public class DateUtils {
 	/**
 	 * 一秒钟的毫秒数
 	 */
-	public static final int ONE_SECOND_MILLISECOND = 1000;
+	public static final long ONE_SECOND_MILLISECOND = 1000;
 
 	/**
 	 * 一分钟的毫秒数
 	 */
-	public static final int ONE_MINUTE_MILLISECOND = 60 * ONE_SECOND_MILLISECOND;
+	public static final long ONE_MINUTE_MILLISECOND = 60 * ONE_SECOND_MILLISECOND;
 
 	/**
 	 * 一小时的毫秒数
 	 */
-	public static final int ONE_HOUR_MILLISECOND = 60 * ONE_MINUTE_MILLISECOND;
+	public static final long ONE_HOUR_MILLISECOND = 60 * ONE_MINUTE_MILLISECOND;
 
 	/**
 	 * 一天的毫秒数
 	 */
-	public static final int ONE_DAY_MILLISECOND = 24 * ONE_HOUR_MILLISECOND;
+	public static final long ONE_DAY_MILLISECOND = 24 * ONE_HOUR_MILLISECOND;
 
 	/**
 	 * 一分钟的秒数
 	 */
-	public static final int ONE_MINUTE_SECOND = ONE_MINUTE_MILLISECOND / ONE_SECOND_MILLISECOND;
+	public static final long ONE_MINUTE_SECOND = ONE_MINUTE_MILLISECOND / ONE_SECOND_MILLISECOND;
 
 	/**
 	 * 1小时的秒数
